Fix #135175: Incorrect curve bounds overlay and no radius
There are two separate but related issues here. First is that the
overlay doesn't properly use all of an object's evaluated geometry
for the bounds and only uses the data that matches the object type.
Second is that the radius isn't included in the bounds for the new
curves type. They are related because using the proper evaluated
geometry would remove the effect of the radius if the second problem
wasn't also fixed.

This commit adds a separate cache for the bounds with the radius, which
is now used by default. The old cache is kept around for backward
compatibility in the bounding box geometry node, where a new
"Use Radius" option accesses the old behavior.