Fix #135134: Tooltip Clip Middle for Library Path
We are showing full paths to the library file on some tooltips, like
for brush assets. We don't wrap these properly (only if it contains
spaces), but we probably wouldn't want to anyway. This PR just clips
the middle of the path (with horizontal ellipsis) when over our maximum
width, favoring the last portion (after last path separator).