Shape Keys: Tweak value slider range behavior
Part of #135095.

The request was to remove hard limits for the slider and have the min
and max controls affect the soft range instead. However, currently there
is no functionality for dynamic soft ranges for RNA properties. While that
could probably be resolved with some work, it may be okay to instead
treat the slider min and max properties as the hard range. That's what
this commit does.