Fix #99026: Enables VSE Channels panel in "Sequence & Preview" view
Addresses https://projects.blender.org/blender/blender/issues/99026

Up until recently, when in VSE "Sequence & Preview" view type, the
"View/Channels" menu option wasn't doing anything when enabled,
not revealing the Channels panel.

The pull request below fixed the issue by disabling the "View/"Channels"
menu option when in the "Sequence & Preview" view type.
https://projects.blender.org/blender/blender/pulls/131368

The intent of this new change is to enable back the "View/Channels"
option and fix the Channels panel rendering inconsistencies when in
"Sequence & Preview" view, which I assume was the reason why it wasn't
used in this view type so far.

The UI fixes include drawing a solid rectangle above the channels list
following the same height of the Sequencer panel, as well as drawing
an additional rectangle to extend the Sequencer scrubbing timeline
element, which adds visual consistency to the design.