UI: Padding When Opening Menu When One is Already Open
When you open a menu or popover, moving your mouse to a neighboring
root menu button will switch to that one. For example when you open the
"File" menu and move to the right to have "Edit" open automatically.
This currently causes problems with some neighboring menus, especially
when they are wide like the area list. This PR adds a safety padding
to the open item. This generally means that once you open a menu, you
will need to move your mouse into the neighbor's content, not just hit
the edge. This still feels natural, but greatly reduces opening the
next menu accidentally.