Fix #135328: Unable to set default Grease Pencil eraser
Incorrectly changed in 9e8c037375

The Grease Pencil `eraser_brush` property is still accessed and set
directly from Python, we cannot remove setting access to the property in
favor of the AssetWeakReference without first making other changes. For
now, this patch re-adds the RNA definitions for the property to make it
editable again.