Fix #127227: Image Editor mask proportional editing has distorted circle
Only the case for non-square images.

In the transform system, `initTransInfo` > `setTransformViewAspect` sets
up the transform aspect. For masks in the Image Editor, it uses
`ED_space_image_get_aspect` [which does not handle non-square images
actually, unlike `ED_space_image_get_uv_aspect` which takes this into
account].

The transform system itself seems to handle all this well, also `t-
>center_global` seems to be correct in context of the transform system.
This all works well for the different modes in the Image Editor, also
UVs in the Image Editor also draw the proportional editing circle
correctly (it is also drawn correctly in e.g. the MCE).

The only context where it fails is the masks in the Image Editor
(drawing the proportional editing circle, where both `GPU_matrix_scale`
with the transform aspect and the `t->center_global` as the drawing
center are wrong).

So to resolve, get an appropriate aspect and position for the drawing
(without touching what the transform system uses internally).