Cleanup: Use Array for field operation inputs instead of Vector
The number of inputs doesn't change after the field operation is
created, so it's a bit simpler to use Array instead. In the few cases
where there are more than 4 inputs for an operation, we can avoid
an extra copy step too. I used NoInitization and placement new for
the generic code path. Since this is so commonly executed it seemed
potentially worth it here.