Split out draw() function into draw & draw_buffer
This is done so any arguments in the future that only make sense
for the buffer/GPU calls don't complicate the function signature
for the other drawing method.

It also means that we can't accidentally attempt to use the GPU
in background mode - if the buffer isn't bound.

The color function now sets the color for both buffer & non buffer
drawing.