Geometry Nodes: Curves to Grease Pencil: Add option to set the depth order
When generating Grease Pencil from scratch, e.g. using the
`Curves to Grease Pencil` node, the depth order always defaults to
"2D layer" ordering. This is not always (maybe even in a majority of
cases the desired behavior.

This adds an option to the `Curves to Grease Pencil` node to set the
depth ordering explicitly.

Resolves #129458.