UI Test: Tablet Pen Tilt Direction
This PR corrects the Windows Ink Y Tilt values (currently using
absolute value so we lose direction), matching direction with those
given by WinTab. This also updates all the comments for tilt to make
it clear what directions expected of the tilt ranges. And this adds
a (for testing purposes only) a display on the Status Bar while drawing
annotation lines to help confirm these directions.