UI: Increase Border Edge Drag Width
When resizing areas by dragging at the edges, the hit area is wider
than the visible border. This extra border padding is extended in this
PR by another one pixel on either side. This means at 1X scale the hit
width is two pixels wider, 4 pixels wider at 2X scale (or on Retina).
This is possible because it adjusts for scrollbars at the edge of the
area.