Fix #135969: Crash exporting grease pencil with no materials as svg/pdf
Having no material is not uncommon, e.g. this can happen in blender when
a Curve is coverted to grease pencil.

If there is no material, we can still regard that stroke as visible, but
skip special material stroke/fill handling. Just export an unfilled,
black stroke in that case.

That still seems to output valid files which can be opened.