Fix: Set more appropriate flags in imbuf on save
The flags are overwritten on save, as the file format may have
changed, but were always set to IB_byte_data. This was noticable in
the interface, where the "Half Float Precision" checkbox should
only be visible for 32bit float images, but showed up after saving
an image as a new half float exr.