Fix: UI tests dependent on mouse position dont work on Mac
Because Retina Macs have a greater than 1 pixel density, tests that
attempt to define a range of positions by enumerating values based on
the `window.width` and `window.height` values does not behave as
expected.

To fix this, use the `pixel_size` property so that window sizes
correspond to expected values based on operating system.