Fix: Use non-zero timer value for UI test event simulation
Previously, the timer was defined to wait a certain number of ticks
while queuing itself immediately. This caused various failures in tests
that use multiple windows, presumably due to changes in different
depsgraphs not propagating before assertions on object states occurred.

Changing this tick replay from using an internal counter to simply
waiting for 1/60th of a second both fixes the issue and results in the
test suite running faster (23 seconds with this patch vs 25 seconds
without).