Fix #135999: Extensions: Preserve full maintainer string on display
Previously when an addon is expanded on the UI, the "Maintainer" string
would be cut short until a `<` character, presumably to extract e-mail,
however it's not using the email in the field, and it's not consistent
between "legacy addons" and "extensions". This fix makes it so that the
full "Maintainer" string is always displayed.