CMake: Add code coverage support for clang
Pretty bare bones but gets the job done, unlike the gcc
tooling, this will work for release builds, the performance cost
of is on the high side of things the full test suite tests take over
an hour for me with code coverage support enabled on a release build
have not timed a debug build. Given developers can just run their
tests to get coverage data over what they are working on, i feel this
is still useful tooling to have.

This adds the same 2 targets as the gcc support does

coverage-reset - this removes the collected code coverage data and report
coverage-report - This merges the collected data and generates the report

Given a browser popping up is undesirable on the CI infra and editing some
cmake code to convince it not to do that, is cumbersome, this does NOT open
a browser like the gcc tooling does for that an extra target has been added

show-coverage-report - This generates the report and opens it up in the browser

This relies on llvm-cov and llvm-profdata being available if not found code
coverage is disabled.

Tested on windows as I currently do not have a working linux environment

The workflow one would have would be

- Build with code coverage enabled
- Run `ctest -R your_test`
- Run `make/ninja show-coverage-report`

Note: A full test run requires an obscene amount of disk space, a complete
test run takes about 125GB and takes 12 minutes to merge, so provision the
COMPILER_CODE_COVERAGE_DATA_DIR folder accordingly