Render: Disable auto texture space normalization by default
Generated texture coordinates by default are normalized to 0..1 in the
bounding box of the mesh. These are most commonly used for procedural
texturing, where this normalization most of the time is not helpful and
caused the texture to non-uniformly stretched along some axes.

Also expand tooltip to clarify what this option does.