UI: Use snapping when drag & dropping collections
Part of #134755 / #134766.

Previously dropping collections would use some point under the mouse cursor, I
think projected based on the 3D cursor location. According to feedback this is
pretty useless and unpredictable, snapping to surfaces under the mouse cursor
is preferred.

With this the collection or collaction instance will be transformed so objects
are placed relative to the dropping point (plus the collection instance offset,
if any).