UI: Move Toolbar Scroll Bars to the Left
Toolbars are aligned to the left sides of their areas and have a
(hidden) right edge that can be dragged to resize or hide it. If there
isn't enough vertical space to show all items then you also have a
scroll bar on the same edge. These two things conflict badly and is
almost impossible to use with a tablet pen. The mouse cursor changes to
indicate dragging over the scrollbar but doesn't work. Dragging the
scollbar requires doing so to the right of it. This PR moves the scroll
bars to the left side where there is already some padding space.