Fix #135039: Crash on startup with Vulkan & LIBDECOR
The placeholder buffer used a SHM buffer, not DMA.
It seems like recent drivers require a DMA buffer as they
run a DMA sync on the buffer - which fails with an error on SHM buffers.

Workaround this by calling swapBuffers when creating the window to
initialize the Vulkan context to remove the need for a dummy buffer.