Sculpt: Reduce common size of BMesh neighbor vector
This commit reduces the default inline allocated size when accessing a
vertex's neighbors in Dyntopo from 64 elements to 16 elements. In most
cases, for a tesselated triangle mesh, a vertex will have 8 neighbors.

To support future tuning of this and avoid having the variable defined
across many files, the constant is extracted into the `sculpt_intern.hh`
header file.