UI: Fix curvemap zoom-out button polling
Sometimes the zoom-out button on curvemaps was shown when it shouldn't
have been -- if "use clipping" is disabled then the user could
erroneously zoom out too far.

Fix by adding some bias. `FLT_EPSILON` is not enough since error is
introduced for every multiply done in a zoom operation.

Also add an explanation for the magic constant that was in
`curvemap_buttons_zoom_in` previously.