Geometry Node: Is Single Value
Node for checking if a `field` is a sparse representation defined by
a single constant value.

Intended to be used for checking status of 'input fields' such as the
Influence/vertex group input for Curve Deform. This node allows adding
conditional execution of those inputs in cases where users don't actually
specify any input and just leaves a default value.