Fix #136327: Deleting group node crashes when there are multiple outputs.
The group output node is updated as part of the topology cache.
When there are multiple group output nodes the first one with an "active"
flag is picked. However, if no group output is flagged this will leave
the group output unchanged in the cache, even if that node has been deleted.

Always initialize the group output runtime cache to null to avoid invalid pointers.