Geometry Nodes: Improve performance with many inputs
Resolves #136183

To avoid quadratic worst case runtime when gathering values from
the modifier properties, build a temporary VectorSet of the modifier's
IDProperties. In the file from #136183, this change improves playback
performance by 20% for me, from 50 to 60ms.

Ideally IDProperty groups would have constant time lookup on their
own, but that's a much larger change, and a smaller change for just
Geometry Nodes is not so invasive.