Simplify constraint solver by combining position and velocity in a single node.
Solving positions and velocities separately was done initially to avoid doing the
velocity estimate inside the constraint solver. But this adds a lot of complexity in
the node setup and isn't really related to "physics" since it's just a kinematic
time derivative.

Now the constraint solver is a single node, which is much more manageable.